Added a TransactionCallbackPayload - felt easier to understand that magic class methods
